Allow plugin versions to specify additional dependencies to be installed for tests
When running in a test environment, any "peerDependencies" specified by a package
are not automatically installed. Additionally, when supporting multiple versions
of the same package, sometimes the required dependency needs to be strictly defined
based on the package version.

This change adds an `additionalDependencies` field to the plugin exports which are
installed during the services installation. These dependencies can be unique for
each version of the package to guarentee version compatibility.