Knex: added error handling support for v0.20.11 and above
v0.20.11 replaced bluebird library with native promises and as a result
async operations raising errors were not caught by the try catch block.
Wrapping the async operations in a new Promise requivalent to the
try..catch block fixed the problem. We now return a new promise that
proxies to the inner promise returned by Knex but catches in errors in
process and records them as tags.