Added `withNonReportingScope()` method to tracer
This method allows users to entirely disable tracing within a given
scope. It accepts a single callback function and ensures any operations
taking places within that callback function will not be traced.

This is useful to disable tracing for certain code paths. For example,
SignalFx tracing/metrics libraries can wrap their network operations
with this method to ensure they don't generate spans for their own
ingest operations.

Implementation:

It introduce the following two new classes:

- NonReportingSpan
- NonReportingSpanContext

The NonReporingSpan acts just like a regular span with the following
diffences:

1. It creates an instance of NonReportingSpanContext instead of SpanContext.
2. It does not report itself to the tracer or recorder ever. This means
   the spans is never recorded and exported by the tracer.

- The newly introduced `tracer.withNonReportingScope()` method creates a new
`NonReportingSpan` instance and sets it as the active span in the scope.
- When the tracer asked to create a new span, it looks at the active
span and if the active span is an instance of the DisbaledSpan, it'll
create a new DisbaledSpan. If the parent is a regular span, then it'll
create a regular span as well preserving current behavior.

This way an enitre code path can be exluded from tracing without having
to modify application or library code executed in that code path.