Prevent span export path from being traced
We use a URL based filter function to avoid tracing outgoing requests to
the the trace export endpoint but this is unreliable as some times the
server can redirect the client to a different URL or port. We already
have built mechanism to prevent tracing certain code paths that we use
in the metrics library. This commit uses the same mechanism to prevent
tracing the span export code path as well.